% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildIsoNet.R
\name{buildIsoNet}
\alias{buildIsoNet}
\title{Title}
\usage{
buildIsoNet(dMDT_network_info, ensembl_data, iso_network, kallisto_counts)
}
\arguments{
\item{dMDT_network_info}{result of switch_calculator}

\item{ensembl_data}{ensembl data}

\item{iso_network}{isoform interaction network}

\item{kallisto_counts}{kallisto count of the network to check if the interaction partners are expressed}
}
\value{
dataframe
}
\description{
Title
}
