% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_csmdts.R
\name{find_csmdts}
\alias{find_csmdts}
\title{MDTs found in \%n of samples}
\usage{
find_csmdts(samples_mdt_files, cutoff)
}
\arguments{
\item{samples_mdt_files}{mdt files that is found by find_mdt() function}

\item{cutoff}{cutoff to define percentages of samples having that MDT in the cohort}
}
\value{
dataframe
}
\description{
MDTs found in \%n of samples
}
