% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_mdts.R
\name{find_mdts}
\alias{find_mdts}
\title{Find MDTs in each sample}
\usage{
find_mdts(data_matrix, cutoff, min_exp)
}
\arguments{
\item{data_matrix}{non-redundant expression data of transcripts}

\item{cutoff}{minimum enrichment value}

\item{min_exp}{minimum cutoff value for transcript expression}
}
\value{
dataframe
}
\description{
Find MDTs in each sample
}
