% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDT_boxplot.R
\name{MDT_boxplot}
\alias{MDT_boxplot}
\title{Plot MDT Distribitions}
\usage{
MDT_boxplot(cond1_MDTs, cond2_MDTs, title, x_axis_label, y_axis_label)
}
\arguments{
\item{cond1_MDTs}{output of find_mdts function for the condition 1}

\item{cond2_MDTs}{output of find_mdts function for the condition 2}
}
\value{
ggplot
}
\description{
Plot MDT Distribitions
}
