% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotExpression.R
\name{plotExpression}
\alias{plotExpression}
\title{Title}
\usage{
plotExpression(dMDT, MDT, sampleName, cond1_exp, cond2_exp)
}
\arguments{
\item{dMDT}{ENST id}

\item{MDT}{ENST id}

\item{sampleName}{sample id}
}
\value{
plot
}
\description{
Title
}
