% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_Seq.R
\name{prepare_Seq}
\alias{prepare_Seq}
\title{Title}
\usage{
prepare_Seq(fastaFile, ENSG_ENST_ENSP)
}
\arguments{
\item{fastaFileDir}{fasta file of ENST sequences}

\item{ENST_geneNamesDir}{ENST Gene Name/ENSG matching}
}
\value{
dataframe
}
\description{
Title
}
