% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_seq_length.R
\name{prepare_seq_length}
\alias{prepare_seq_length}
\title{Calculate lengths of Transcripts' peptide sequences if exists}
\usage{
prepare_seq_length(fastaFileDir)
}
\arguments{
\item{fastaFileDir}{fasta file of ENST sequences}
}
\value{
dataframe
}
\description{
Calculate lengths of Transcripts' peptide sequences if exists
}
