% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_redundants.R
\name{remove_redundants}
\alias{remove_redundants}
\title{Remove Redundant Transcripts}
\usage{
remove_redundants(
  data_matrix,
  enst_ensp_sequences,
  mane_select = NULL,
  peptide_lengths = NULL
)
}
\arguments{
\item{data_matrix}{Bulk RNA Seq expression matrix from samples, columns are samples and first row is gene ids, second row is transcript ids and the rest of the rows are samples}

\item{enst_ensp_sequences}{amino acid sequence of each ensp}

\item{mane_select}{mane transcript list is needed to decide which transcript will be kept}
}
\value{
dataframe
}
\description{
Remove Redundant Transcripts
}
