% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_calculator.R
\name{switch_calculator}
\alias{switch_calculator}
\title{Find Isoform Switching Events}
\usage{
switch_calculator(
  cancer_ori,
  gtex_ori,
  exp_values_pcawg,
  exp_values_gtex,
  ensp_sequences,
  cutoff,
  cutoff_rate,
  cutoff_other_MDTs,
  ensp_only
)
}
\arguments{
\item{cancer_ori}{MDT files of cancer cohort}

\item{gtex_ori}{MDT files of normal cohort}

\item{exp_values_pcawg}{abundance dataframe of cancer samples}

\item{exp_values_gtex}{abundance dataframe of gtex samples}

\item{ensp_sequences}{ensp sequences of transcripts}

\item{cutoff}{cutoff}

\item{cutoff_rate}{enrichment value (for example MDT1 should be expressed >=2 than the next one)}

\item{cutoff_other_MDTs}{cutoff to decide percentage of MDTs found in normal samples}

\item{ensp_only}{transcripts having ENSP ids}
}
\value{
datatable
}
\description{
Find Isoform Switching Events
}
