% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_network.R
\name{visualize_network}
\alias{visualize_network}
\title{Title}
\usage{
visualize_network(protein, iso_network)
}
\arguments{
\item{protein}{ENSP id of protein of interest}

\item{iso_network}{isoform interaction network}
}
\value{
plot
}
\description{
Title
}
